feat: Created a new fragment for input and calculation value of convert.

+++ app/src/main/java/com/cobaltumapps/simplecalculator/v15/converter/calculator/ConverterDiagonalMatrixCalculator.kt
@@ -21,14 +21,13 @@ class ConverterDiagonalMatrixCalculator {
                     val result = BigDecimal(userInput).multiply(
                         valueToConvert.toBigDecimal(), MathContext.DECIMAL128)
 
-                    resultArray.add(result)
+                    resultArray.add(result.toDouble())
 
                 } catch (ex: ArrayIndexOutOfBoundsException) {
                     return arrayOf()
                 }
 
             }
-
         }
         return resultArray.toTypedArray()
     }
